Update schema loader to use groovy wrapper
Since the schema loader code is in groovy, we use the
java wrapper to invoke it, instead of having separate
java code.
